// WYNKWeb/src/app/Models/OfferTran.ts
export class OfferTran {

  OfferType: string;
  OnePlusOfferValue: number;
  OFID: number;
  LMID: number;
  LTID: number;
  DiscountPercentage: number;
  DiscountAmount: number;
  UOMID: number;
  GSTPercentage: number;
  GSTTaxValue: number;
  CGSTPercentage: number;
  CGSTTaxValue: number;
  SGSTPercentage: number;
  IGSTPercentage: number;
  TotalIGSTTaxValue: number;
  CESSPercentage: number;
  AdditionalCESSPercentage: number;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;

}